Batch 30 fixes: eliminate all Wconversion/format/unused warnings in kernel core
Eliminates 15 pre-existing warnings across 4 kernel source files:

kernel.c:
  - Cast strlen() return to uint32_t in vfs_write calls (2 sites)

gdt.c:
  - Explicit uint8_t cast for gdt base_high assignment

fault.c:
  - Remove unused 'current_sz' variable in stack-grow path
  - Fix format specifiers: %lx -> %llx with unsigned long long casts

syscall.c:
  - Cast need_end to uint32_t for vfs_read calls (2 sites: read + fd_read)
  - Cast len to uint16_t for net_link_send (2nd param is uint16_t)
  - Cast revents to int16_t in poll/ppoll syscalls (4 sites)
  - Change uint32_t saved_off -> uint64_t to match pfd->offset type (2 sites)
  - Remove unused 'dirp' variable in getdents64

Build now has ZERO C warnings (only NASM dword-overflow in .asm files remain).

+++ src/kernel/gdt.c
@@ -29,7 +29,7 @@ static uint8_t __attribute__((aligned(16))) early_ist1_stack[EARLY_IST1_STACK_SI
 static void gdt_set_entry(int i, uint32_t base, uint32_t limit, uint8_t access, uint8_t gran) {
     gdt[i].base_low    = base & 0xFFFF;
     gdt[i].base_mid    = (base >> 16) & 0xFF;
-    gdt[i].base_high   = (base >> 24) & 0xFF;
+    gdt[i].base_high   = (uint8_t)((base >> 24) & 0xFF);
     gdt[i].limit_low   = limit & 0xFFFF;
     gdt[i].granularity  = ((limit >> 16) & 0x0F) | (gran & 0xF0);
     gdt[i].access      = access;
